+++ pages/_campus/edit.vue
@@ -134,6 +134,39 @@
           </div>
         </div>
         <ec-field
+          label="Distance à laquelle le soutenu peut réserver une course"
+          field-id="reservation-scope"
+        >
+          <div class="columns">
+            <div class="column select">
+              <select
+                id="select-scope"
+                @change="defineScope"
+              >
+                <option value="">
+                  Choisissez une unité :
+                </option>
+                <option value="hour">
+                  Heures
+                </option>
+                <option value="day">
+                  Jours
+                </option>
+              </select>
+            </div>
+            <div class="column">
+              <input
+                id="reservation-scope"
+                v-model="localReservationScope"
+                class="input"
+                type="number"
+                :disabled="!selectScope"
+                :placeholder="!selectScope && 'Veuillez sélectionner une unité'"
+              >
+            </div>
+          </div>
+        </ec-field>
+        <ec-field
           label="Catégories"
           field-id="categories"
         >
@@ -200,7 +233,7 @@ import weekdaysSelect from '~/components/form/weekdays.vue';
 import rideDuration from '~/components/form/ride-duration.vue';
 
 const EDITABLE_FIELDS = 'id,name,location,phone(drivers,everybody),categories(id,label),'
-    + 'information,timezone,workedDays,workedHours,defaultRideDuration';
+    + 'information,timezone,workedDays,workedHours,defaultRideDuration,defaultReservationScope';
 export default {
   components: {
     ecField,
@@ -221,6 +254,33 @@ export default {
       },
     };
   },
+  data() {
+    return {
+      selectScope: null,
+    };
+  },
+  computed: {
+    localReservationScope: {
+      get() {
+        if (this.selectScope === 'day') {
+          return Math.floor(this.campus.defaultReservationScope / 3600 / 24);
+        }
+        if (this.selectScope === 'hour') {
+          return Math.floor(this.campus.defaultReservationScope / 3600);
+        }
+        return null;
+      },
+      set(value) {
+        if (this.selectScope === 'day') {
+          this.campus.defaultReservationScope = value * 3600 * 24;
+        }
+        if (this.selectScope === 'hour') {
+          this.campus.defaultReservationScope = value * 3600;
+        }
+        return value;
+      },
+    },
+  },
   methods: {
     async edit(campus) {
       let data = {};
@@ -237,6 +297,9 @@ export default {
         params: { id: data.id },
       });
     },
+    defineScope({ target }) {
+      this.selectScope = target.value;
+    },
   },
 };
 </script>
@@ -249,5 +312,9 @@ export default {
     padding: $size-small/2 $size-small;
     margin: 0;
     border-radius: $radius-small;
+  }
+
+  .select select {
+    width: 100%;
   }
 </style>