+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add explicit error messages on 
 * Add possibility for admin to edit his own campus
 * Round percents in dashboard
+* Add custom reservation scope for campus
 ### Fix
 * Updated login page ui
 * Changed dropdown title when creating a new car in planning